requirements.txt
Flask==2.0.2
psycopg2==2.9.1
python-dotenv==0.19.1
gunicorn==20.1.0
